requirements.txt
Django==5.2.12
djangorestframework==3.17.1
django-extensions==4.1
mysqlclient==2.2.7
python-dotenv==1.1.1
requests==2.33.1
requests-oauthlib==2.0.0
Sphinx==8.2.3
